clean up logs and comments
