+++ src/components/Workspace/WorkspaceEditor.tsx
@@ -22,7 +22,7 @@ import { useNetworkSummaryStore } from '../../store/NetworkSummaryStore'
 import { NdexNetworkSummary } from '../../models/NetworkSummaryModel'
 import { AppConfigContext } from '../../AppConfigContext'
 import { Workspace } from '../../models/WorkspaceModel'
-import { putNetworkViewToDb, getUiStateFromDb } from '../../store/persist/db'
+import { putNetworkViewToDb } from '../../store/persist/db'
 import { NetworkView } from '../../models/ViewModel'
 import { useWorkspaceManager } from '../../store/hooks/useWorkspaceManager'
 
@@ -31,7 +31,7 @@ import { SnackbarMessageList } from '../Messages'
 import { NetworkBrowserPanel } from './NetworkBrowserPanel/NetworkBrowserPanel'
 
 import { SidePanel } from './SidePanel/SidePanel'
-import { DEFAULT_UI_STATE, useUiStateStore } from '../../store/UiStateStore'
+import { useUiStateStore } from '../../store/UiStateStore'
 import { Ui } from '../../models/UiModel'
 import { PanelState } from '../../models/UiModel/PanelState'
 import { OpenRightPanelButton } from './SidePanel/OpenRightPanelButton'
@@ -87,7 +87,6 @@ const WorkSpaceEditor = (): JSX.Element => {
   const ui: Ui = useUiStateStore((state) => state.ui)
   const setPanelState: (panel: Panel, panelState: PanelState) => void =
     useUiStateStore((state) => state.setPanelState)
-  const setUi = useUiStateStore((state) => state.setUi)
 
   const { panels, activeNetworkView } = ui
 
@@ -260,18 +259,6 @@ const WorkSpaceEditor = (): JSX.Element => {
     exclusiveSelect(currentNetworkId, selectedNodes, selectedEdges)
   }
 
-  const loadUiState = (): void => {
-    void getUiStateFromDb().then((uiState) => {
-      if (uiState !== undefined) {
-        setUi(uiState)
-        // console.log('loaded ui from db', uiState)
-      } else {
-        // console.log('setting default')
-        setUi(DEFAULT_UI_STATE)
-      }
-    })
-  }
-
   /**
    * Initializations
    */
@@ -282,7 +269,6 @@ const WorkSpaceEditor = (): JSX.Element => {
     window.addEventListener('resize', windowWidthListener)
 
     restorePanelStates()
-    loadUiState()
 
     return () => {
       window.removeEventListener('resize', windowWidthListener)
